[sim] Abstract hardware factory
