+++ Blocks/src/sim/java/com/google/blocks/ftcrobotcontroller/hardware/HardwareItemMap.java
@@ -16,6 +16,8 @@
 
 package com.google.blocks.ftcrobotcontroller.hardware;
 
+import com.dekaresearch.robotcore.simulation.SimulationConstants;
+import com.dekaresearch.simulation.hardwarefactory.SimulationHardwareFactory;
 import com.qualcomm.ftccommon.configuration.RobotConfigFile;
 import com.qualcomm.ftccommon.configuration.RobotConfigFileManager;
 import com.qualcomm.robotcore.exception.RobotCoreException;
@@ -30,8 +32,6 @@ import com.qualcomm.robotcore.hardware.configuration.MotorControllerConfiguratio
 import com.qualcomm.robotcore.hardware.configuration.ReadXMLFileHandler;
 import com.qualcomm.robotcore.hardware.configuration.ServoControllerConfiguration;
 import com.qualcomm.robotcore.util.RobotLog;
-import com.dekaresearch.robotcore.simulation.SimulationConstants;
-import com.dekaresearch.simulation.romi.RomiHardwareFactory;
 
 import org.xmlpull.v1.XmlPullParser;
 
@@ -41,8 +41,8 @@ import java.util.Collections;
 import java.util.Comparator;
 import java.util.HashSet;
 import java.util.List;
-import java.util.SortedMap;
 import java.util.Set;
+import java.util.SortedMap;
 import java.util.TreeMap;
 
 /**
@@ -61,7 +61,7 @@ public class HardwareItemMap {
    */
   public static HardwareItemMap newHardwareItemMap() {
     if(SimulationConstants.isSimulation) {
-      return new HardwareItemMap(RomiHardwareFactory.createHardwareMap(null));
+      return new HardwareItemMap(SimulationHardwareFactory.createHardwareMap(null));
     }
 
     try {
